début de correction pour la conversion %p, mais casse tout ?

+++ ft_printf/ft_dectohex.h
@@ -6,14 +6,15 @@
 /*   By: cvillene <cvillene@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/07 07:16:53 by cvillene          #+#    #+#             */
-/*   Updated: 2025/11/07 08:39:35 by cvillene         ###   ########.fr       */
+/*   Updated: 2025/11/13 17:29:01 by cvillene         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_DECTOHEX_H
 # define FT_DECTOHEX_H
 # include <stdlib.h>
+# include <limits.h>
 
-char	*ft_dectohex(int n, int islower);
+char	*ft_dectohex(long long n, int islower);
 
 #endif
